Add organization and plugin name to required info
Fixes #28.
Add the name and the organization of the plugin to the required info.
As the name and the organization is needed for all plugin the
implementation of this is done in the Device class.
This way, only that information that is different for that plugin is
added in the plugin.